Monster House

+++ Assets/Scripts/Dungeon/Structure/Cell.cs
@@ -4,7 +4,7 @@ public class Cell
 {
     public bool traversed = false;
     public CellType cellType;
-    public float roomNum;
+    public int roomNum;
     public int x;
     public int y;
     public GameObject itemOnCell;
